refactor: all ducks to be plural
