+++ ui/src/duck/index.ts
@@ -2,21 +2,21 @@ import { HYDRATE } from 'next-redux-wrapper';
 import { combineReducers, Reducer } from 'redux';
 
 import auth from './auth';
-import designation from './designation';
-import employee from './employee';
+import designations from './designations';
+import employees from './employees';
 import flashMessages from './flashMessages';
-import organization from './organizations';
-import project from './project';
-import team from './team';
+import organizations from './organizations';
+import projects from './projects';
+import teams from './teams';
 
 const reducer = combineReducers({
   [auth.name]: auth.reducer,
-  [designation.name]: designation.reducer,
+  [designations.name]: designations.reducer,
   [flashMessages.name]: flashMessages.reducer,
-  [employee.name]: employee.reducer,
-  [organization.name]: organization.reducer,
-  [project.name]: project.reducer,
-  [team.name]: team.reducer,
+  [employees.name]: employees.reducer,
+  [organizations.name]: organizations.reducer,
+  [projects.name]: projects.reducer,
+  [teams.name]: teams.reducer,
 });
 
 const rootReducer: Reducer<RootState> = (state, action): RootState => {
